9283: PR Comments
Story: AB#9283
Task: AB#9642
