+++ src/NHSD.BuyingCatalogue.Ordering.Api/Startup.cs
@@ -93,7 +93,7 @@ namespace NHSD.BuyingCatalogue.Ordering.Api
                 .AddTransient<ICreatePurchasingDocumentService, CreatePurchasingDocumentService>()
                 .AddTransient<ICreateOrderItemValidator, OrderItemValidator>()
                 .AddTransient<IUpdateOrderItemValidator, OrderItemValidator>()
-                .AddTransient<IAttachmentCsvWriter<PatientNumbersPriceType>, AttachmentCsvWriter<PatientNumbersPriceType, PatientNumbersPriceTypeMap>>();
+                .AddTransient<IAttachmentCsvWriter<PatientNumbersPriceType>, CsvStreamWriter<PatientNumbersPriceType, PatientNumbersPriceTypeMap>>();
 
             services.RegisterHealthChecks(connectionString, smtpSettings);
