+++ tests/NHSD.BuyingCatalogue.Ordering.Api.IntegrationTests/Steps/Support/ScenarioContextKeys.cs
@@ -6,6 +6,5 @@
         internal const string OrganisationPayloadDictionary = nameof(OrganisationPayloadDictionary);
         internal const string SupplierAddress = nameof(SupplierAddress);
         internal const string SupplierContact = nameof(SupplierContact);
-        internal const string OrderId = nameof(OrderId);
     }
 }
